+++ packages/ccextractor/src/index.js
@@ -1,2 +1,0 @@
-export { CCExtractorAppliance } from './CCExtractorAppliance'
-export { CCExtractorLine } from './CCExtractorLine'