+++ packages/video-caption-extractor/src/index.js
@@ -1,0 +1,2 @@
+export { VideoCaptionExtractorAppliance } from './VideoCaptionExtractorAppliance'
+export { CCExtractorLine } from './CCExtractorLine'